requirements.txt
cffi==1.17.1
cryptography==43.0.1
psutil==6.0.0
pycparser==2.22
pycryptodome==3.21.0
Flask==3.0.3
